requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.0
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
protobuf==3.8.0
PyJWT==1.4.2
pytz==2019.1
six==1.12.0
Werkzeug==0.15.4
